separate some methods from CommonUtils to io.FileUtils
move to FileUtils
